data-ingestion/requirements.txt
# Data Ingestion Dependencies

# HTTP & API Clients
httpx==0.26.0
aiohttp==3.9.1
sodapy==2.2.0  # CDC Socrata API

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Geospatial
h3==3.7.6
shapely==2.0.2

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
geoalchemy2==0.14.3

# GCP
google-cloud-functions==1.13.0
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.14.0

# Logging
structlog==24.1.0

# Validation
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
